+++ src/main/java/com/followMe/delivery_server/delivery/application/DeliveryService.java
@@ -3,17 +3,16 @@ package com.followMe.delivery_server.delivery.application;
 import com.followMe.common.pagination.PageRequest;
 import com.followMe.common.pagination.PageResponse;
 import com.followMe.delivery_server.delivery.application.dto.DeliveryCreateResponse;
-import com.followMe.delivery_server.delivery.application.dto.DeliverySearchCondition;
+import com.followMe.delivery_server.delivery.application.dto.DeliveryRequest;
+import com.followMe.delivery_server.delivery.application.dto.DeliveryResponse;
 import com.followMe.delivery_server.delivery.domain.Delivery;
+import com.followMe.delivery_server.delivery.domain.DeliverySearchCondition;
 import com.followMe.delivery_server.delivery.domain.Node;
 import com.followMe.delivery_server.delivery.domain.UserContext;
 import com.followMe.delivery_server.delivery.domain.exception.DeliveryNotFoundException;
-import com.followMe.delivery_server.delivery.domain.query.DeliveryQueryPort;
 import com.followMe.delivery_server.delivery.domain.repository.DeliveryRepository;
 import com.followMe.delivery_server.delivery.domain.service.DeliveryPermissionChecker;
 import com.followMe.delivery_server.delivery.domain.service.HubRouteInfo;
-import com.followMe.delivery_server.delivery.presentation.dto.DeliveryRequest;
-import com.followMe.delivery_server.delivery.presentation.dto.DeliveryResponse;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -39,9 +38,10 @@ public class DeliveryService {
 
   @Transactional(readOnly = true)
   public DeliveryResponse.Detail getDelivery(UserContext user, UUID deliveryId) {
-    DeliveryResponse.Detail detail = deliveryQueryPort.findDeliveryById(deliveryId);
-    permissionChecker.checkReadAccess(user, detail.shipments());
-    return detail;
+    Delivery delivery =
+        deliveryRepository.findById(deliveryId).orElseThrow(DeliveryNotFoundException::new);
+    permissionChecker.checkReadAccess(user, delivery.getShipments());
+    return deliveryQueryPort.findDeliveryById(deliveryId);
   }
 
   @Transactional(readOnly = true)
@@ -53,9 +53,10 @@ public class DeliveryService {
 
   @Transactional(readOnly = true)
   public DeliveryResponse.Detail getDeliveryByOrderId(UserContext user, UUID orderId) {
-    DeliveryResponse.Detail detail = deliveryQueryPort.findDeliveryByOrderId(orderId);
-    permissionChecker.checkReadAccess(user, detail.shipments());
-    return detail;
+    Delivery delivery =
+        deliveryRepository.findByOrderId(orderId).orElseThrow(DeliveryNotFoundException::new);
+    permissionChecker.checkReadAccess(user, delivery.getShipments());
+    return deliveryQueryPort.findDeliveryByOrderId(orderId);
   }
 
   @Transactional
