+++ src/main/java/com/followMe/delivery_server/delivery/application/dto/DeliveryRequest.java
@@ -1,4 +1,4 @@
-package com.followMe.delivery_server.delivery.presentation.dto;
+package com.followMe.delivery_server.delivery.application.dto;
 
 import java.util.UUID;
