+++ src/main/java/com/followMe/delivery_server/delivery/domain/service/DeliveryPermissionChecker.java
@@ -1,15 +1,14 @@
 package com.followMe.delivery_server.delivery.domain.service;
 
-import com.followMe.delivery_server.delivery.application.dto.DeliverySearchCondition;
 import com.followMe.delivery_server.delivery.domain.Delivery;
+import com.followMe.delivery_server.delivery.domain.DeliverySearchCondition;
 import com.followMe.delivery_server.delivery.domain.Shipment;
 import com.followMe.delivery_server.delivery.domain.UserContext;
-import com.followMe.delivery_server.delivery.presentation.dto.ShipmentResponse;
 import java.util.List;
 
 public interface DeliveryPermissionChecker {
 
-  void checkReadAccess(UserContext user, List<ShipmentResponse.Detail> shipments);
+  void checkReadAccess(UserContext user, List<Shipment> shipments);
 
   void checkCancelAccess(UserContext user, Delivery delivery);
