+++ src/main/java/com/followMe/delivery_server/delivery/presentation/ShipmentController.java
@@ -1,12 +1,11 @@
 package com.followMe.delivery_server.delivery.presentation;
 
 import com.followMe.common.response.ApiResponse;
-import com.followMe.delivery_server.delivery.application.ShipmentServiceImpl;
-import com.followMe.delivery_server.delivery.application.dto.DeliveryResponseDto;
-import com.followMe.delivery_server.delivery.application.dto.ShipmentStatusUpdateRequest;
-import jakarta.validation.Valid;
+import com.followMe.delivery_server.delivery.application.dto.ShipmentRequest;
+import com.followMe.delivery_server.delivery.application.dto.ShipmentResponse;
 import com.followMe.delivery_server.delivery.domain.UserContext;
 import com.followMe.delivery_server.delivery.domain.UserRole;
+import jakarta.validation.Valid;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
