--- src/main/java/com/followMe/delivery_server/delivery/presentation/dto/ShipmentRequest.java
@@ -1,7 +1,0 @@
-package com.followMe.delivery_server.delivery.presentation.dto;
-
-import com.followMe.delivery_server.delivery.domain.enums.ShipmentStatus;
-
-public class ShipmentRequest {
-  public record UpdateStatus(ShipmentStatus status) {}
-}
